requirements.txt
pyyaml>=6.0
requests>=2.28.0
flask>=2.2.0